requirements.txt
beautifulsoup4==4.12.3
requests
lxml
selenium
python_dotenv
fuzzywuzzy